Organize symbols
