+++ src/app/exercise/exercises/utility/settings/NumberOfSegmentsSetting.ts
@@ -1,4 +1,4 @@
-import { Exercise } from '../../Exercise';
+import { Exercise } from '../../../Exercise';
 
 export type NumberOfSegmentsSetting = {
   numberOfSegments: number;
