+++ src/app/shared/ts-utility/Constructor.ts
@@ -1,0 +1,1 @@
+export type Constructor<G = any> = { new(...args: any[]): G };
